requirements.txt
Pillow~=10.1.0
piexif~=1.1.3